pyproject.toml
[tool.poetry]
name = "task-api"
version = "0.1.0"
description = "Create task and other basic schemas"
authors = ["PatLabCorp <patrick.alonso.fc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.9.1"
mypy = "^1.11.2"
pytest = "^8.3.3"
pre-commit = "^3.8.0"
gunicorn = "^23.0.0"
flask = "^3.0.3"
flask-cors = "^5.0.0"
prometheus-client = "^0.20.0"
pymongo = "^4.8.0"
python-magic = "^0.4.27"
sonyflake-py = "^1.3.0"
nanoid = "^2.0.0"
pandas = "^2.2.2"
pika = "^1.3.2"
grpcio-tools = "^1.66.1"
python-dotenv = "^1.0.1"
grpcio-reflection = "^1.66.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
